# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Dokumentasi REST API backend AgriPadi untuk autentikasi, profil, upload gambar, percakapan, dan alur diagnosis hama padi.",
        "title": "AgriPadi Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/auth/login": {
            "post": {
                "description": "Login menggunakan nomor telepon dan password untuk mendapatkan JWT.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Login pengguna",
                "parameters": [
                    {
                        "description": "Data login",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.AuthResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengakhiri sesi pengguna di sisi aplikasi.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Logout pengguna",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.LogoutResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/register": {
            "post": {
                "description": "Membuat akun pengguna baru untuk aplikasi AgriPadi.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Registrasi pengguna",
                "parameters": [
                    {
                        "description": "Data registrasi",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/auth.AuthResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/conversations/{conversation_id}/messages": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengambil seluruh pesan pada satu percakapan milik pengguna.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Messages"
                ],
                "summary": "Daftar pesan percakapan",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID percakapan",
                        "name": "conversation_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.MessageListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.APIErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.APIErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengirim pesan pada percakapan. Payload dapat berisi teks, lampiran gambar, dan kandidat deteksi CNN untuk memulai alur diagnosis.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Messages"
                ],
                "summary": "Kirim pesan pengguna",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID percakapan",
                        "name": "conversation_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Data pesan",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.OrchestratorResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.APIErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.APIErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/profile": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengambil detail profil pengguna yang sedang login.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profile"
                ],
                "summary": "Ambil profil pengguna",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/profile.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Memperbarui nama, email, atau nomor telepon pengguna.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profile"
                ],
                "summary": "Perbarui profil pengguna",
                "parameters": [
                    {
                        "description": "Data profil",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/profile.UpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/profile.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/profile/password": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengubah password akun pengguna yang sedang login.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profile"
                ],
                "summary": "Ubah password",
                "parameters": [
                    {
                        "description": "Data perubahan password",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/profile.ChangePasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "auth.AuthResponse": {
            "type": "object",
            "properties": {
                "expires_at": {
                    "type": "string",
                    "example": "2023-10-27T10:00:00Z"
                },
                "token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
                },
                "token_type": {
                    "type": "string",
                    "example": "Bearer"
                },
                "user": {
                    "$ref": "#/definitions/auth.UserResponse"
                }
            }
        },
        "auth.LoginRequest": {
            "type": "object",
            "required": [
                "password",
                "phone_number"
            ],
            "properties": {
                "password": {
                    "type": "string",
                    "example": "katasandi123"
                },
                "phone_number": {
                    "type": "string",
                    "maxLength": 32,
                    "example": "081234567890"
                }
            }
        },
        "auth.LogoutResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "auth.RegisterRequest": {
            "type": "object",
            "required": [
                "confirm_password",
                "email",
                "name",
                "password",
                "phone_number"
            ],
            "properties": {
                "confirm_password": {
                    "type": "string",
                    "example": "katasandi123"
                },
                "email": {
                    "type": "string",
                    "example": "[EMAIL_ADDRESS]"
                },
                "name": {
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 1,
                    "example": "Budi Santoso"
                },
                "password": {
                    "type": "string",
                    "minLength": 8,
                    "example": "katasandi123"
                },
                "phone_number": {
                    "type": "string",
                    "maxLength": 32,
                    "example": "081234567890"
                }
            }
        },
        "auth.UserResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2023-10-27T10:00:00Z"
                },
                "email": {
                    "type": "string",
                    "example": "[EMAIL_ADDRESS]"
                },
                "id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                },
                "name": {
                    "type": "string",
                    "example": "Budi Santoso"
                },
                "phone_number": {
                    "type": "string",
                    "example": "081234567890"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2023-10-27T10:00:00Z"
                }
            }
        },
...
        "domain.MessageType": {
            "type": "string",
            "enum": [
                "text",
                "image"
            ],
            "x-enum-varnames": [
                "MessageTypeText",
                "MessageTypeImage"
            ]
        },
...
        "dto.APIErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "unauthorized"
                }
            }
        },
...
        "dto.ChatAction": {
            "type": "object",
            "properties": {
                "label": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/dto.ActionType"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "dto.ConversationMode": {
            "type": "string",
            "enum": [
                "consultation",
                "diagnose"
            ],
            "x-enum-varnames": [
                "ConversationModeConsultation",
                "ConversationModeDiagnose"
            ]
        },
        "dto.CreateAttachmentRequest": {
            "type": "object",
            "required": [
                "mime_type",
                "size",
                "type",
                "url"
            ],
            "properties": {
                "checksum": {
                    "type": "string"
                },
                "height": {
                    "type": "integer"
                },
                "mime_type": {
                    "type": "string"
                },
                "size": {
                    "type": "integer",
                    "minimum": 1
                },
                "thumbnail_url": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "width": {
                    "type": "integer"
                }
            }
        },
        "dto.CreateMessageRequest": {
            "type": "object",
            "required": [
                "content",
                "type"
            ],
            "properties": {
                "attachments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.CreateAttachmentRequest"
                    }
                },
                "content": {
                    "type": "string"
                },
                "conversation_id": {
                    "type": "string"
                },
                "detections": {
                    "description": "top-k detections from cnn model",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.DetectionCandidateRequest"
                    }
                },
                "type": {
                    "$ref": "#/definitions/domain.MessageType"
                }
            }
        },
        "dto.DetectionCandidateRequest": {
            "type": "object",
            "required": [
                "confidence",
                "label",
                "model"
            ],
            "properties": {
                "confidence": {
                    "type": "number",
                    "maximum": 1,
                    "minimum": 0
                },
                "label": {
                    "type": "string"
                },
                "latency_ms": {
                    "type": "integer"
                },
                "model": {
                    "type": "string"
                },
                "rank": {
                    "type": "integer"
                },
                "source": {
                    "type": "string"
                }
            }
        },
...
        "dto.DiagnoseFlowState": {
            "type": "string",
            "enum": [
                "idle",
                "image_analyzed",
                "collect_symptoms",
                "normalize_symptoms",
                "collect_severity",
                "collect_growth_stage",
                "generate_result",
                "completed"
            ],
            "x-enum-varnames": [
                "DiagnoseFlowStateIdle",
                "DiagnoseFlowStateImageAnalyzed",
                "DiagnoseFlowStateCollectSymptoms",
                "DiagnoseFlowStateNormalizeSymptoms",
                "DiagnoseFlowStateCollectSeverity",
                "DiagnoseFlowStateCollectGrowthStage",
                "DiagnoseFlowStateGenerateResult",
                "DiagnoseFlowStateCompleted"
            ]
        },
...
        "dto.MessageListResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.MessageResponse"
                    }
                }
            }
        },
        "dto.MessageResponse": {
            "type": "object",
            "properties": {
                "actions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ChatAction"
                    }
                },
                "ai": {
                    "$ref": "#/definitions/dto.AIMessageMetadata"
                },
                "attachments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.AttachmentResponse"
                    }
                },
                "content": {
                    "type": "string"
                },
                "conversation_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_streaming": {
                    "type": "boolean"
                },
                "payload": {
                    "$ref": "#/definitions/dto.OrchestratorPayload"
                },
                "role": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "dto.OrchestratorPayload": {
            "type": "object",
            "properties": {
                "diagnose": {
                    "$ref": "#/definitions/dto.DiagnosePayload"
                }
            }
        },
        "dto.OrchestratorResponse": {
            "type": "object",
            "properties": {
                "actions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ChatAction"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "mode": {
                    "$ref": "#/definitions/dto.ConversationMode"
                },
                "payload": {
                    "$ref": "#/definitions/dto.OrchestratorPayload"
                },
                "session_id": {
                    "type": "string"
                },
                "state": {
                    "$ref": "#/definitions/dto.DiagnoseFlowState"
                }
            }
        },
...
        "profile.ChangePasswordRequest": {
            "type": "object",
            "required": [
                "confirm_password",
                "current_password",
                "new_password"
            ],
            "properties": {
                "confirm_password": {
                    "type": "string"
                },
                "current_password": {
                    "type": "string"
                },
                "new_password": {
                    "type": "string",
                    "minLength": 8
                }
            }
        },
        "profile.Response": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "profile.UpdateRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 1
                },
                "phone_number": {
                    "type": "string",
                    "maxLength": 32
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Masukkan token JWT dengan format: Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}